Time: 0 ms (100.00%) | Memory: 8.5 MB (91.41%) - LeetSync

+++ 9-palindrome-number/palindrome-number.cpp
@@ -10,6 +10,6 @@ public:
             rev=rev*10+temp%10;
             temp/=10;
         }
-        return rev==x;
+        return (int)rev==x;
     }
 };
